moved BoundingBoxDataType into common
