+++ main/src/main/java/de/wintervillage/main/plot/listener/PlayerInteractListener.java
@@ -2,7 +2,7 @@ package de.wintervillage.main.plot.listener;
 
 import de.wintervillage.common.paper.util.BoundingBox2D;
 import de.wintervillage.main.WinterVillage;
-import de.wintervillage.main.persistent.BoundingBoxDataType;
+import de.wintervillage.common.paper.persistent.BoundingBoxDataType;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
 import org.bukkit.entity.Player;
